+++ start_multimodal_services.sh
@@ -12,5 +12,5 @@ CONFIG_FILE_PATH="./configs/example.configuration_multimodal_bm_body_tracking_st
 echo "--------------------"
 CONFIG_FILE_PATH="./configs/example.configuration_multimodal_bm_body_tracking_streaming.json" docker compose up inference-data-processing-body-tracking -d
 echo "--------------------"
-CONFIG_FILE_PATH="./configs/example.configuration_multimodal_bm_body_tracking_streaming.json" docker compose up fusion-layer -d
+#CONFIG_FILE_PATH="./configs/example.configuration_multimodal_bm_body_tracking_streaming.json" docker compose up fusion-layer -d
 docker compose up mock-xroom-writer -d